SM: Input field functions moved to AppInput.js
